+++ CHANGELOG.md
@@ -8,7 +8,8 @@
     - Send ping if we haven't received any data from the server in 30 seconds.
     - Show S3RV3R NOT R3SPOND1NG if the server hasn't responded in 45 seconds. (15 seconds after ping)
     - Close connection and try to reconnect if the server hasn't responded in 90 seconds. (60 seconds after ping)
-    
+ - Fallback to PyQt5. (for Windows 7 users mainly)
+ 
 ### Fixed
  - Error when manually moving group.
  - EOF on send not resulting in a disconnect.